requirements.txt

bayespy==0.5.6
Django==1.11
django-dag==1.4.2
django-picklefield==1.0.0
graphviz==0.6
Pillow==4.2.1
